+++ src/components/myPage/DecoProfile.tsx
@@ -1,7 +1,7 @@
 import MyBorder from './MyBorder';
 import { useState } from 'react';
 import MyInvenAward from './MyInvenAward';
-import { Deco } from './Deco.styles';
+import { D } from './Styled.MyPage/Deco.styles';
 import * as S from '../../pages/Shop.style';
 import { styled } from 'styled-components';
 
@@ -19,59 +19,33 @@ const DecoProfile = () => {
     }
   };
   return (
-    <DecoContainer>
-      <DecoTitle>프로필 꾸미기</DecoTitle>
+    <D.Page>
+      <D.Title>프로필 꾸미기</D.Title>
 
-      <Container>
-        <Deco.ButtonContainer>
-          <S.Button
-            onClick={() => setSelectedDecoMenu('Border')}
-            style={{
-              backgroundColor:
-                selectedDecoMenu === 'Border' ? '#FF96DB' : '#FFEBF7',
-              color: selectedDecoMenu === 'Border' ? '#ffffff' : 'black',
-            }}
-          >
-            테두리
-          </S.Button>
-          <S.Button
-            onClick={() => setSelectedDecoMenu('Award')}
-            style={{
-              backgroundColor:
-                selectedDecoMenu === 'Award' ? '#FF96DB' : '#FFEBF7',
-              color: selectedDecoMenu === 'Award' ? '#ffffff' : 'black',
-            }}
-          >
-            칭호
-          </S.Button>
-        </Deco.ButtonContainer>
-        <div>{renderDecoComponent()}</div>
-      </Container>
-    </DecoContainer>
+      <D.ButtonContainer>
+        <S.Button
+          onClick={() => setSelectedDecoMenu('Border')}
+          style={{
+            backgroundColor:
+              selectedDecoMenu === 'Border' ? '#FF96DB' : '#FFEBF7',
+            color: selectedDecoMenu === 'Border' ? '#ffffff' : 'black',
+          }}
+        >
+          테두리
+        </S.Button>
+        <S.Button
+          onClick={() => setSelectedDecoMenu('Award')}
+          style={{
+            backgroundColor:
+              selectedDecoMenu === 'Award' ? '#FF96DB' : '#FFEBF7',
+            color: selectedDecoMenu === 'Award' ? '#ffffff' : 'black',
+          }}
+        >
+          칭호
+        </S.Button>
+      </D.ButtonContainer>
+      <D.Container>{renderDecoComponent()}</D.Container>
+    </D.Page>
   );
 };
 export default DecoProfile;
-const DecoContainer = styled.div`
-  position: absolute;
-  top: 170px;
-`;
-const DecoTitle = styled.div`
-  position: absolute;
-  top: -25px;
-  left: 150px;
-  width: 200px;
-  height: 32px;
-  color: #000;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  letter-spacing: -0.36px;
-`;
-const Container = styled.div`
-  display: grid;
-  position: absolute;
-  flex-direction: column;
-  margin-top: -18%;
-  margin-left: 150px;
-`;
