+++ src/components/myPage/LikedAnime.tsx
@@ -10,7 +10,7 @@ import { styled } from 'styled-components';
 import viewDetail from '../../assets/viewdetail.svg';
 import useViewport from '../../hooks/useViewPort';
 import LikeSvg from '../anime-recommend/LikeSvg';
-import LikedSkeleton from './LikedSkeleton';
+import LikedSkeleton from './Skeleton.MyPage/LikedSkeleton';
 import goShop from '../../assets/goShop.png';
 import { PaginationTwo } from '../PagenationTwo';
 
@@ -116,10 +116,10 @@ const LikedAnime = () => {
   };
   const likedList =
     Array.isArray(filteredLiked) && filteredLiked.length > 0 ? (
-      <GridContainer>
+      <L.GridContainer>
         {displayedAnime.map((filteredLiked, index) => (
           <div key={index}>
-            <Liked.Container
+            <L.Container
               onClick={() => navigate(`/recommend/${filteredLiked.anime_id}`)}
             >
               <PosterImage
@@ -134,11 +134,11 @@ const LikedAnime = () => {
                     : undefined
                 }
               />
-              <AnimeTitle>
+              <L.AnimeTitle>
                 {animeTitles[filteredLiked.anime_id] &&
                   animeTitles[filteredLiked.anime_id].name}
-              </AnimeTitle>
-              <HoverContent>
+              </L.AnimeTitle>
+              <L.HoverContent>
                 <HoverViewDetail>
                   <p>자세히 보기</p>
                   <img
@@ -161,45 +161,45 @@ const LikedAnime = () => {
                         </HoveredAnimeGenreTag>
                       ))}
                 </HoveredAnimeGenre>
-                <LikedAnimeGenre>
+                <L.LikedAnimeGenre>
                   {animeTitles[filteredLiked.anime_id] &&
                     animeTitles[filteredLiked.anime_id]?.genres
                       ?.slice(0, 2)
                       .map((genre, index) => (
-                        <GenreTag key={index}>#{genre}</GenreTag>
+                        <L.GenreTag key={index}>#{genre}</L.GenreTag>
                       ))}
-                </LikedAnimeGenre>
-              </HoverContent>
-            </Liked.Container>
+                </L.LikedAnimeGenre>
+              </L.HoverContent>
+            </L.Container>
           </div>
         ))}
-      </GridContainer>
+      </L.GridContainer>
     ) : (
-      <GoAnimeContainer>
-        <GoAnimeMessage>좋아요를 누른 애니메이션이 없어요!</GoAnimeMessage>
-        <GoAnimeRecommend
+      <L.NoContainer>
+        <L.NoMessage>좋아요를 누른 애니메이션이 없어요!</L.NoMessage>
+        <L.GoRecommend
           onClick={() => {
             navigate('/recommend');
           }}
         >
           애니메이션 추천받으러 가기
           <img src={goShop} alt="고샾" />
-        </GoAnimeRecommend>
-      </GoAnimeContainer>
+        </L.GoRecommend>
+      </L.NoContainer>
     );
   return (
-    <LikedContainer>
+    <L.LikedContainer>
       {user && topTags.length > 0 && (
-        <TopTags>
-          {user?.nickname}님은 <Tags>#{topTags.join('#')}</Tags>을 좋아해요!
-        </TopTags>
+        <L.TopTags>
+          {user?.nickname}님은 <L.Tags>#{topTags.join('#')}</L.Tags>을 좋아해요!
+        </L.TopTags>
       )}
-      <LikedTitle>찜한 목록</LikedTitle>
+      <L.Title>찜한 목록</L.Title>
 
-      <FullPage>
-        <LikedList>{likedList}</LikedList>
+      <L.FullPage>
+        <L.List>{likedList}</L.List>
         {Array.isArray(displayedAnime) && displayedAnime.length >= 0 && (
-          <Page>
+          <L.Page>
             <PaginationTwo
               currentPage={currentPage}
               totalPages={totalPages}
@@ -207,89 +207,15 @@ const LikedAnime = () => {
               isPreviousDisabled={currentPage === 1}
               isNextDisabled={currentPage >= totalPages}
             />
-          </Page>
+          </L.Page>
         )}
-      </FullPage>
-    </LikedContainer>
+      </L.FullPage>
+    </L.LikedContainer>
   );
 };
 export default LikedAnime;
-const LikedContainer = styled.div`
-  margin-top: 10px;
-`;
 
-const Page = styled.div`
-  position: absolute;
-  margin-right: -680px;
-  margin-top: -330px;
-`;
-const TopTags = styled.div`
-  border-radius: 999px;
-  position: absolute;
-  background: var(--main-light-2, #f3e7ff);
-  width: auto;
-  display: flex;
-  height: 32px;
-  padding: 8px 20px;
-  justify-content: space-between;
-  align-items: center;
-  gap: 8px;
-  margin-top: -360px;
-  margin-left: 130px;
-`;
-const FullPage = styled.div`
-  position: absolute;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-left: 150px;
-`;
-const Tags = styled.div`
-  font-weight: 600;
-`;
-const LikedList = styled.div`
-  margin-top: -280px;
-  position: absolute;
-  left: 10px;
-`;
-const LikedTitle = styled.div`
-  position: relative;
-  margin-left: 150px;
-  top: -300px;
-  width: 200px;
-  height: 32px;
-  color: #000;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  letter-spacing: -0.36px;
-`;
-const GridContainer = styled.div`
-  display: grid;
-  gap: 0px 20px;
-  grid-template-columns: repeat(3, 1fr);
-`;
-const AnimeTitle = styled.div`
-  width: 220px;
-  height: 19px;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 19px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  margin: 10px;
-`;
-const HoverContent = styled.div`
-  align-items: center;
-  justify-content: center;
-  display: grid;
-  margin-left: 110px;
-`;
-const HoverViewDetail = styled.div`
+export const HoverViewDetail = styled.div`
   visibility: hidden;
   display: flex;
   margin-top: -250px;
@@ -306,54 +232,12 @@ const HoverViewDetail = styled.div`
     margin-left: 12px;
   }
 `;
-const LikedAnimeGenre = styled.div`
-  height: 16px;
-  display: flex;
-  justify-content: flex-start;
-  margin-left: -110px;
-  margin-bottom: 24px;
-  margin-top: -12px;
-  padding: 5px;
-  gap: 4px;
-`;
-const GenreTag = styled.div`
-  justify-content: center;
-  align-items: center;
-  padding: 4px 12px;
-  height: 16px;
-  background: #efefef;
-  border-radius: 999px;
-`;
-const HoveredAnimeGenre = styled.div`
-  visibility: hidden;
-  display: flex;
-  align-items: flex-start;
-  margin-left: -110px;
-  padding: 5px;
-  position: absolute;
-  top: 0;
-  left: 50;
-`;
-const HoveredAnimeGenreTag = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 150px;
-  height: 16px;
-  background: #f3e7ff;
-  padding: 7px;
-  border-radius: 3px;
-  width: 80px;
-  height: auto;
-  border-radius: 50px;
-  text-align: center;
-`;
-const PosterImage = styled.img`
+export const PosterImage = styled.img`
   width: 280px;
   height: 175px;
   border-radius: 10px;
 `;
-const LikedInfoTitle = styled.div`
+export const LikedInfoTitle = styled.div`
   visibility: hidden;
   width: 100%;
   color: white;
@@ -371,8 +255,157 @@ const LikedInfoTitle = styled.div`
   transform: translate(-45%, -90%);
   overflow: hidden;
 `;
+export const HoveredAnimeGenreTag = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 150px;
+  height: 16px;
+  background: #f3e7ff;
+  padding: 7px;
+  border-radius: 3px;
+  width: 80px;
+  height: auto;
+  border-radius: 50px;
+  text-align: center;
+`;
+export const HoveredAnimeGenre = styled.div`
+  visibility: hidden;
+  display: flex;
+  align-items: flex-start;
+  margin-left: -110px;
+  padding: 5px;
+  position: absolute;
+  top: 0;
+  left: 50;
+`;
 
-const Liked = {
+const L = {
+  LikedContainer: styled.div`
+    margin-top: 10px;
+  `,
+  GridContainer: styled.div`
+    display: grid;
+    gap: 0px 20px;
+    grid-template-columns: repeat(3, 1fr);
+  `,
+  Title: styled.div`
+    position: relative;
+    margin-left: 150px;
+    top: -300px;
+    width: 200px;
+    height: 32px;
+    color: #000;
+    font-size: 24px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+    letter-spacing: -0.36px;
+  `,
+  TopTags: styled.div`
+    border-radius: 999px;
+    position: absolute;
+    background: var(--main-light-2, #f3e7ff);
+    width: auto;
+    display: flex;
+    height: 32px;
+    padding: 8px 20px;
+    justify-content: space-between;
+    align-items: center;
+    gap: 8px;
+    margin-top: -360px;
+    margin-left: 130px;
+  `,
+
+  NoContainer: styled.div`
+    display: grid;
+    align-items: center;
+
+    justify-content: center;
+    margin-left: 190%;
+    margin-top: 85%;
+  `,
+  GoRecommend: styled.button`
+    background-color: #8200ff;
+    color: #fff;
+    width: 226.5px;
+    height: 48px;
+    border-radius: 30px;
+    border-color: transparent;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 10px;
+    cursor: pointer;
+  `,
+  NoMessage: styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 10px;
+  `,
+
+  Page: styled.div`
+    position: absolute;
+    margin-right: -680px;
+    margin-top: -330px;
+  `,
+
+  FullPage: styled.div`
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: flex-end;
+    margin-left: 150px;
+  `,
+  Tags: styled.div`
+    font-weight: 600;
+  `,
+  List: styled.div`
+    margin-top: -280px;
+    position: absolute;
+    left: 10px;
+  `,
+
+  AnimeTitle: styled.div`
+    width: 220px;
+    height: 19px;
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 19px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    margin: 10px;
+  `,
+  HoverContent: styled.div`
+    align-items: center;
+    justify-content: center;
+    display: grid;
+    margin-left: 110px;
+  `,
+
+  LikedAnimeGenre: styled.div`
+    height: 16px;
+    display: flex;
+    justify-content: flex-start;
+    margin-left: -110px;
+    margin-bottom: 24px;
+    margin-top: -12px;
+    padding: 5px;
+    gap: 4px;
+  `,
+  GenreTag: styled.div`
+    justify-content: center;
+    align-items: center;
+    padding: 4px 12px;
+    height: 16px;
+    background: #efefef;
+    border-radius: 999px;
+  `,
+
   Container: styled.div`
     position: relative;
     display: flex;
@@ -431,31 +464,3 @@ const Liked = {
     }
   `,
 };
-const GoAnimeRecommend = styled.button`
-  background-color: #8200ff;
-  color: #fff;
-  width: 226.5px;
-  height: 48px;
-  border-radius: 30px;
-  border-color: transparent;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px;
-  cursor: pointer;
-`;
-const GoAnimeMessage = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px;
-`;
-const GoAnimeContainer = styled.div`
-  display: grid;
-  align-items: center;
-
-  justify-content: center;
-  margin-left: 190%;
-  margin-top: 85%;
-`;
