+++ src/components/myPage/MyProfile.tsx
@@ -1,16 +1,17 @@
 import { useAtomValue } from 'jotai'; // Import from jotai
-import { Profile } from './MyPage.styles';
+import { Profile } from './Styled.MyPage/MyPage.styles';
 import MyProfileAward from './MyProfileAward';
 import MyPoint from './MyPoint';
 import * as userStore from '../../store/userStore';
 import ProfileWithBorder from '../ProfileWithBorder';
 import useViewport from '../../hooks/useViewPort';
+import Loading from '../Loading/Loading';
 const MyProfile = () => {
   const user = useAtomValue(userStore.user);
   const { width } = useViewport();
 
   if (!user) {
-    return <div>로딩중...</div>;
+    return <Loading />;
   }
 
   return (
