+++ src/components/myPage/Styled.MyPage/Deco.styles.ts
@@ -1,16 +1,75 @@
 import styled from 'styled-components';
-export const Deco = {
+export const D = {
+  Page: styled.div`
+    position: absolute;
+    top: 170px;
+  `,
+  Title: styled.div`
+    position: absolute;
+    top: -25px;
+    left: 150px;
+    width: 200px;
+    height: 32px;
+    color: #000;
+    font-size: 24px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+    letter-spacing: -0.36px;
+  `,
+  Container: styled.div`
+    display: grid;
+    position: absolute;
+    flex-direction: column;
+    margin-top: -18%;
+    margin-left: 150px;
+  `,
   ButtonContainer: styled.div`
     display: flex;
   `,
-  Point: styled.div`
-    background-color: #f3e7ff;
-    border: none;
-    border-radius: 10px;
-    padding: 8px 20px;
-  `,
 };
 export const B = {
+  Container: styled.div`
+    position: absolute;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 5px 1px;
+  `,
+  BorderContainer: styled.div`
+    position: absolute;
+  `,
+  BorderImg: styled.img`
+    width: 150px;
+    height: auto;
+    object-fit: cover;
+    margin: 20px;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.1);
+  `,
+  ButtonContainer: styled.div`
+    justify-content: space-between;
+    width: 100px;
+    align-items: center;
+    margin-left: 25px;
+  `,
+  BorderName: styled.div`
+    width: 100%;
+    height: 20%;
+    background-color: white;
+    margin-top: 5px;
+    line-height: 25px;
+  `,
+  BorderPage: styled.div`
+    position: absolute;
+    justify-content: center;
+    top: -45px;
+    left: 810px;
+  `,
+
+  BorderLoading: styled.div`
+    margin-left: 500px;
+  `,
   NoneContainer: styled.div`
     display: grid;
     align-items: center;
@@ -25,21 +84,11 @@ export const B = {
     object-fit: cover;
     margin: 10px;
   `,
-  BorderImg: styled.img`
-    width: 150px;
-    height: auto;
-    object-fit: cover;
-
-    margin: 20px;
-
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.1);
-  `,
-  Number: styled.div`
-    width: 100%;
-    height: 80%;
-    background-size: cover;
+  NoneMessage: styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 10px;
   `,
   NoneButton: styled.div`
     background-color: #8200ff;
@@ -53,17 +102,7 @@ export const B = {
     margin: 10px;
     cursor: pointer;
   `,
-  NoneMessage: styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 10px;
-  `,
-  Border: styled.img`
-    width: 100%;
-    height: auto;
-    object-fit: cover;
-  `,
+
   Equip: styled.button<{ is_equipped: boolean }>`
     position: RELATIVE;
     font-size: 14px;
@@ -73,7 +112,7 @@ export const B = {
     width: 80px;
     height: 30px;
     background-color: ${(props) => (props.is_equipped ? '#F3E7FF' : 'white')};
-color:black;
+    color:black;
     border-radius: 6px;
     border: ${(props) =>
       props.is_equipped ? '1px solid #c88fff' : '1px solid #d9d9d9'};
@@ -90,33 +129,19 @@ color:black;
         cursor: not-allowed;
       }
   `,
-  BorderContainer: styled.div`
-    width: 268px;
-    height: 300px;
-    gap: 0px;
-  `,
-  ButtonContainer: styled.div`
-    justify-content: space-between;
-    width: 100px;
-    align-items: center;
-    margin-left: 25px;
-  `,
-
-  Container: styled.div`
-    position: absolute;
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 5px 1px;
-  `,
-  BorderName: styled.div`
-    width: 100%;
-    height: 20%;
-    background-color: white;
-    margin-top: 5px;
-    line-height: 25px;
-  `,
 };
 export const A = {
+  Container: styled.div`
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 50px;
+    padding: 10px;
+  `,
+
+  Loading: styled.div`
+    margin-left: 500px;
+  `,
+
   NoneMessage: styled.div`
     display: flex;
     justify-content: center;
@@ -149,26 +174,13 @@ export const A = {
     object-fit: cover;
     margin: 10px;
   `,
-  Name: styled.div`
-    background-color: #efefef;
-    padding: 8px;
-    border-radius: 999px;
-    text-align: center;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    letter-spacing: -0.3px;
-    display: flex;
-    height: 48px;
-    width: 260px;
-    padding: 12px;
+  Page: styled.div`
+    position: absolute;
     justify-content: center;
-    align-items: center;
-    gap: 8px;
-    align-self: stretch;
-    margin: 10px;
+    top: -565px;
+    left: 630px;
   `,
+
   Equip: styled.button<{ is_equipped: boolean }>`
     width: 80px;
     height: 30px;
@@ -191,8 +203,5 @@ export const A = {
     &:disabled:hover {
       cursor: not-allowed;
     }
-  `,
-  Container: styled.div`
-    position: absolute;
   `,
 };
